Made use of unsigned int in Dim consistent, fixed warnings
Former-commit-id: 30a4c6c3e693621398b6a4ba78f66b7e53ff4a1c

+++ examples/encdec.cc
@@ -48,10 +48,10 @@ struct EncoderDecoder {
       fwd_enc_builder(LAYERS, INPUT_DIM, HIDDEN_DIM, &model) {
 
 
-    p_ie2h = model.add_parameters({long(HIDDEN_DIM * LAYERS * 1.5), long(HIDDEN_DIM * LAYERS * 2)});
-    p_bie = model.add_parameters({long(HIDDEN_DIM * LAYERS * 1.5)});
-    p_h2oe = model.add_parameters({long(HIDDEN_DIM * LAYERS), long(HIDDEN_DIM * LAYERS * 1.5)});
-    p_boe = model.add_parameters({long(HIDDEN_DIM * LAYERS)});
+    p_ie2h = model.add_parameters({unsigned(HIDDEN_DIM * LAYERS * 1.5), unsigned(HIDDEN_DIM * LAYERS * 2)});
+    p_bie = model.add_parameters({unsigned(HIDDEN_DIM * LAYERS * 1.5)});
+    p_h2oe = model.add_parameters({unsigned(HIDDEN_DIM * LAYERS), unsigned(HIDDEN_DIM * LAYERS * 1.5)});
+    p_boe = model.add_parameters({unsigned(HIDDEN_DIM * LAYERS)});
     p_c = model.add_lookup_parameters(INPUT_VOCAB_SIZE, {INPUT_DIM}); 
     p_ec = model.add_lookup_parameters(INPUT_VOCAB_SIZE, {INPUT_DIM}); 
     p_R = model.add_parameters({OUTPUT_VOCAB_SIZE, HIDDEN_DIM});
@@ -91,12 +91,12 @@ struct EncoderDecoder {
     Expression i_nc = i_boe + i_h2oe * i_h;
     
     vector<Expression> oein1, oein2, oein;
-    for (int i = 0; i < LAYERS; ++i) {
+    for (unsigned i = 0; i < LAYERS; ++i) {
       oein1.push_back(pickrange(i_nc, i * HIDDEN_DIM, (i + 1) * HIDDEN_DIM));
       oein2.push_back(tanh(oein1[i]));
     }
-    for (int i = 0; i < LAYERS; ++i) oein.push_back(oein1[i]);
-    for (int i = 0; i < LAYERS; ++i) oein.push_back(oein2[i]);
+    for (unsigned i = 0; i < LAYERS; ++i) oein.push_back(oein1[i]);
+    for (unsigned i = 0; i < LAYERS; ++i) oein.push_back(oein2[i]);
 
     dec_builder.new_graph(cg);
     dec_builder.start_new_sequence(oein);
